+++ test/IntegrationTests/MySqlConnectorTests.cs
@@ -1,0 +1,51 @@
+// <copyright file="MySqlConnectorTests.cs" company="OpenTelemetry Authors">
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+using IntegrationTests.Helpers;
+using Xunit.Abstractions;
+
+namespace IntegrationTests;
+
+[Collection(MySqlCollection.Name)]
+public class MySqlConnectorTests : TestHelper
+{
+    private readonly MySqlFixture _mySql;
+
+    public MySqlConnectorTests(ITestOutputHelper output, MySqlFixture mySql)
+        : base("MySqlConnector", output)
+    {
+        _mySql = mySql;
+    }
+
+    [Theory]
+    [Trait("Category", "EndToEnd")]
+    [Trait("Containers", "Linux")]
+    [MemberData(nameof(LibraryVersion.MySqlConnector), MemberType = typeof(LibraryVersion))]
+    public void SubmitsTraces(string packageVersion)
+    {
+        using var collector = new MockSpansCollector(Output);
+        SetExporter(collector);
+        collector.Expect("MySqlConnector");
+
+        RunTestApplication(new()
+        {
+            Arguments = $"--mysql {_mySql.Port}",
+            PackageVersion = packageVersion
+        });
+
+        collector.AssertExpectations();
+    }
+}